add MegaPi / Laserbot changes based on renebohne/grbl-laserbot

+++ grbl/spindle_control.c
@@ -70,7 +70,7 @@ void spindle_stop()
 
 // Sets spindle speed PWM output and enable pin, if configured. Called by spindle_set_state()
 // and stepper ISR. Keep routine small and efficient.
-void spindle_set_speed(uint16_t pwm_value)
+void spindle_set_speed(uint8_t pwm_value)
 {
   SPINDLE_OCR_REGISTER = pwm_value; // Set PWM output level.
   #ifdef SPINDLE_ENABLE_OFF_WITH_ZERO_SPEED
@@ -97,9 +97,9 @@ void spindle_set_speed(uint16_t pwm_value)
 #ifdef ENABLE_PIECEWISE_LINEAR_SPINDLE
 
   // Called by spindle_set_state() and step segment generator. Keep routine small and efficient.
-  uint16_t spindle_compute_pwm_value(float rpm) // 328p PWM register is 8-bit.
+  uint8_t spindle_compute_pwm_value(float rpm) // Mega2560 Timer2 PWM register is 8-bit.
   {
-    uint16_t pwm_value;
+    uint8_t pwm_value;
     rpm *= (0.010*sys.spindle_speed_ovr); // Scale by spindle speed override value.
     // Calculate PWM register value based on rpm max/min settings and programmed rpm.
     if ((settings.rpm_min >= settings.rpm_max) || (rpm >= RPM_MAX)) {
@@ -140,9 +140,9 @@ void spindle_set_speed(uint16_t pwm_value)
 #else 
 
   // Called by spindle_set_state() and step segment generator. Keep routine small and efficient.
-  uint16_t spindle_compute_pwm_value(float rpm) // Mega2560 PWM register is 16-bit.
+  uint8_t spindle_compute_pwm_value(float rpm) // Mega2560 Timer2 PWM register is 8-bit.
   {
-	uint16_t pwm_value;
+	uint8_t pwm_value;
 	rpm *= (0.010*sys.spindle_speed_ovr); // Scale by spindle speed override value.
 	// Calculate PWM register value based on rpm max/min settings and programmed rpm.
 	if ((settings.rpm_min >= settings.rpm_max) || (rpm >= settings.rpm_max)) {
